+++ poulpy-cpu-avx/src/core_impl.rs
@@ -4,15 +4,7 @@ use poulpy_cpu_ref::ring::CpuRing;
 use crate::{FFT64Avx, NTT4x30Avx};
 #[cfg(feature = "enable-rayon")]
 use crate::{FFT64AvxRayon, NTT4x30AvxRayon};
-use poulpy_core::{
-    impl_conversion_reference_full, impl_decryption_reference_full, impl_encryption_reference_full,
-    impl_gglwe_automorphism_reference_full, impl_gglwe_external_product_reference_full, impl_gglwe_keyswitch_reference_full,
-    impl_gglwe_product_digits_strided_reference, impl_ggsw_automorphism_reference_full,
-    impl_ggsw_external_product_reference_full, impl_ggsw_keyswitch_reference_full, impl_glwe_automorphism_reference_full,
-    impl_glwe_external_product_reference_full, impl_glwe_keyswitch_reference_full, impl_glwe_packing_reference_full,
-    impl_glwe_tensoring_reference, impl_glwe_trace_reference_full, impl_linear_transformation_reference_full,
-    impl_lwe_keyswitch_reference_full,
-};
+use poulpy_core::{impl_gglwe_product_digits_strided_reference, impl_glwe_tensoring_reference};
 use poulpy_hal::layouts::{Module, ScratchArena, VecZnxDftBackendMut, VecZnxDftBackendRef, VmpPMatBackendRef};
 
 impl_glwe_tensoring_reference!(FFT64Avx);
@@ -82,153 +74,15 @@ unsafe impl<R: CpuRing> poulpy_core::oep::GGLWEProductDigitsStridedImpl for NTT4
     }
 }
 
-impl_glwe_automorphism_reference_full!(FFT64Avx);
-impl_glwe_automorphism_reference_full!(crate::FFT64CIAvx);
-impl_glwe_automorphism_reference_full!(NTT4x30Avx);
-impl_glwe_automorphism_reference_full!(crate::NTT4x30CIAvx);
-
-impl_ggsw_automorphism_reference_full!(FFT64Avx);
-impl_ggsw_automorphism_reference_full!(crate::FFT64CIAvx);
-impl_ggsw_automorphism_reference_full!(NTT4x30Avx);
-impl_ggsw_automorphism_reference_full!(crate::NTT4x30CIAvx);
-impl_gglwe_automorphism_reference_full!(FFT64Avx);
-impl_gglwe_automorphism_reference_full!(crate::FFT64CIAvx);
-impl_gglwe_automorphism_reference_full!(NTT4x30Avx);
-impl_gglwe_automorphism_reference_full!(crate::NTT4x30CIAvx);
-
-impl_decryption_reference_full!(FFT64Avx);
-impl_decryption_reference_full!(crate::FFT64CIAvx);
-impl_decryption_reference_full!(NTT4x30Avx);
-impl_decryption_reference_full!(crate::NTT4x30CIAvx);
-impl_glwe_trace_reference_full!(FFT64Avx);
-impl_glwe_trace_reference_full!(crate::FFT64CIAvx);
-impl_glwe_trace_reference_full!(NTT4x30Avx);
-impl_glwe_trace_reference_full!(crate::NTT4x30CIAvx);
-impl_glwe_packing_reference_full!(FFT64Avx);
-impl_glwe_packing_reference_full!(crate::FFT64CIAvx);
-impl_glwe_packing_reference_full!(NTT4x30Avx);
-impl_glwe_packing_reference_full!(crate::NTT4x30CIAvx);
-
-impl_conversion_reference_full!(FFT64Avx);
-impl_conversion_reference_full!(crate::FFT64CIAvx);
-impl_conversion_reference_full!(NTT4x30Avx);
-impl_conversion_reference_full!(crate::NTT4x30CIAvx);
-
-impl_glwe_keyswitch_reference_full!(FFT64Avx);
-impl_glwe_keyswitch_reference_full!(crate::FFT64CIAvx);
-impl_glwe_keyswitch_reference_full!(NTT4x30Avx);
-impl_glwe_keyswitch_reference_full!(crate::NTT4x30CIAvx);
-impl_gglwe_keyswitch_reference_full!(FFT64Avx);
-impl_gglwe_keyswitch_reference_full!(crate::FFT64CIAvx);
-impl_gglwe_keyswitch_reference_full!(NTT4x30Avx);
-impl_gglwe_keyswitch_reference_full!(crate::NTT4x30CIAvx);
-impl_ggsw_keyswitch_reference_full!(FFT64Avx);
-impl_ggsw_keyswitch_reference_full!(crate::FFT64CIAvx);
-impl_ggsw_keyswitch_reference_full!(NTT4x30Avx);
-impl_ggsw_keyswitch_reference_full!(crate::NTT4x30CIAvx);
-impl_lwe_keyswitch_reference_full!(FFT64Avx);
-impl_lwe_keyswitch_reference_full!(crate::FFT64CIAvx);
-impl_lwe_keyswitch_reference_full!(NTT4x30Avx);
-impl_lwe_keyswitch_reference_full!(crate::NTT4x30CIAvx);
-
-impl_encryption_reference_full!(FFT64Avx);
-impl_encryption_reference_full!(crate::FFT64CIAvx);
-poulpy_cpu_ref::impl_sampling_host!(FFT64Avx, fft64);
-poulpy_cpu_ref::impl_sampling_host!(crate::FFT64CIAvx, fft64);
-impl_encryption_reference_full!(NTT4x30Avx);
-impl_encryption_reference_full!(crate::NTT4x30CIAvx);
-poulpy_cpu_ref::impl_sampling_host!(NTT4x30Avx, ntt4x30);
-poulpy_cpu_ref::impl_sampling_host!(crate::NTT4x30CIAvx, ntt4x30);
-
-impl_glwe_external_product_reference_full!(FFT64Avx);
-impl_glwe_external_product_reference_full!(crate::FFT64CIAvx);
-impl_glwe_external_product_reference_full!(NTT4x30Avx);
-impl_glwe_external_product_reference_full!(crate::NTT4x30CIAvx);
-impl_gglwe_external_product_reference_full!(FFT64Avx);
-impl_gglwe_external_product_reference_full!(crate::FFT64CIAvx);
-impl_gglwe_external_product_reference_full!(NTT4x30Avx);
-impl_gglwe_external_product_reference_full!(crate::NTT4x30CIAvx);
-impl_ggsw_external_product_reference_full!(FFT64Avx);
-impl_ggsw_external_product_reference_full!(crate::FFT64CIAvx);
-impl_ggsw_external_product_reference_full!(NTT4x30Avx);
-impl_ggsw_external_product_reference_full!(crate::NTT4x30CIAvx);
-
-impl_linear_transformation_reference_full!(FFT64Avx);
-impl_linear_transformation_reference_full!(crate::FFT64CIAvx);
-impl_linear_transformation_reference_full!(NTT4x30Avx);
-impl_linear_transformation_reference_full!(crate::NTT4x30CIAvx);
-
+poulpy_cpu_ref::impl_cpu_core_defaults!(crate::FFT64Avx, fft64);
+poulpy_cpu_ref::impl_cpu_core_defaults!(crate::FFT64CIAvx, fft64);
+poulpy_cpu_ref::impl_cpu_core_defaults!(crate::NTT4x30Avx, ntt4x30);
+poulpy_cpu_ref::impl_cpu_core_defaults!(crate::NTT4x30CIAvx, ntt4x30);
 #[cfg(feature = "enable-rayon")]
-mod rayon_defaults {
-    use super::*;
-
-    impl_glwe_automorphism_reference_full!(FFT64AvxRayon);
-    impl_glwe_automorphism_reference_full!(crate::FFT64CIAvxRayon);
-    impl_ggsw_automorphism_reference_full!(FFT64AvxRayon);
-    impl_ggsw_automorphism_reference_full!(crate::FFT64CIAvxRayon);
-    impl_gglwe_automorphism_reference_full!(FFT64AvxRayon);
-    impl_gglwe_automorphism_reference_full!(crate::FFT64CIAvxRayon);
-    impl_decryption_reference_full!(FFT64AvxRayon);
-    impl_decryption_reference_full!(crate::FFT64CIAvxRayon);
-    impl_glwe_trace_reference_full!(FFT64AvxRayon);
-    impl_glwe_trace_reference_full!(crate::FFT64CIAvxRayon);
-    impl_glwe_packing_reference_full!(FFT64AvxRayon);
-    impl_glwe_packing_reference_full!(crate::FFT64CIAvxRayon);
-    impl_conversion_reference_full!(FFT64AvxRayon);
-    impl_conversion_reference_full!(crate::FFT64CIAvxRayon);
-    impl_glwe_keyswitch_reference_full!(FFT64AvxRayon);
-    impl_glwe_keyswitch_reference_full!(crate::FFT64CIAvxRayon);
-    impl_gglwe_keyswitch_reference_full!(FFT64AvxRayon);
-    impl_gglwe_keyswitch_reference_full!(crate::FFT64CIAvxRayon);
-    impl_ggsw_keyswitch_reference_full!(FFT64AvxRayon);
-    impl_ggsw_keyswitch_reference_full!(crate::FFT64CIAvxRayon);
-    impl_lwe_keyswitch_reference_full!(FFT64AvxRayon);
-    impl_lwe_keyswitch_reference_full!(crate::FFT64CIAvxRayon);
-    impl_encryption_reference_full!(FFT64AvxRayon);
-    impl_encryption_reference_full!(crate::FFT64CIAvxRayon);
-    poulpy_cpu_ref::impl_sampling_host!(FFT64AvxRayon, fft64);
-    poulpy_cpu_ref::impl_sampling_host!(crate::FFT64CIAvxRayon, fft64);
-    impl_glwe_external_product_reference_full!(FFT64AvxRayon);
-    impl_glwe_external_product_reference_full!(crate::FFT64CIAvxRayon);
-    impl_gglwe_external_product_reference_full!(FFT64AvxRayon);
-    impl_gglwe_external_product_reference_full!(crate::FFT64CIAvxRayon);
-    impl_ggsw_external_product_reference_full!(FFT64AvxRayon);
-    impl_ggsw_external_product_reference_full!(crate::FFT64CIAvxRayon);
-    impl_linear_transformation_reference_full!(FFT64AvxRayon);
-    impl_linear_transformation_reference_full!(crate::FFT64CIAvxRayon);
-
-    impl_glwe_automorphism_reference_full!(NTT4x30AvxRayon);
-    impl_glwe_automorphism_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_ggsw_automorphism_reference_full!(NTT4x30AvxRayon);
-    impl_ggsw_automorphism_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_gglwe_automorphism_reference_full!(NTT4x30AvxRayon);
-    impl_gglwe_automorphism_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_decryption_reference_full!(NTT4x30AvxRayon);
-    impl_decryption_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_glwe_trace_reference_full!(NTT4x30AvxRayon);
-    impl_glwe_trace_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_glwe_packing_reference_full!(NTT4x30AvxRayon);
-    impl_glwe_packing_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_conversion_reference_full!(NTT4x30AvxRayon);
-    impl_conversion_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_glwe_keyswitch_reference_full!(NTT4x30AvxRayon);
-    impl_glwe_keyswitch_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_gglwe_keyswitch_reference_full!(NTT4x30AvxRayon);
-    impl_gglwe_keyswitch_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_ggsw_keyswitch_reference_full!(NTT4x30AvxRayon);
-    impl_ggsw_keyswitch_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_lwe_keyswitch_reference_full!(NTT4x30AvxRayon);
-    impl_lwe_keyswitch_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_encryption_reference_full!(NTT4x30AvxRayon);
-    impl_encryption_reference_full!(crate::NTT4x30CIAvxRayon);
-    poulpy_cpu_ref::impl_sampling_host!(NTT4x30AvxRayon, ntt4x30);
-    poulpy_cpu_ref::impl_sampling_host!(crate::NTT4x30CIAvxRayon, ntt4x30);
-    impl_glwe_external_product_reference_full!(NTT4x30AvxRayon);
-    impl_glwe_external_product_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_gglwe_external_product_reference_full!(NTT4x30AvxRayon);
-    impl_gglwe_external_product_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_ggsw_external_product_reference_full!(NTT4x30AvxRayon);
-    impl_ggsw_external_product_reference_full!(crate::NTT4x30CIAvxRayon);
-    impl_linear_transformation_reference_full!(NTT4x30AvxRayon);
-    impl_linear_transformation_reference_full!(crate::NTT4x30CIAvxRayon);
-}
+poulpy_cpu_ref::impl_cpu_core_defaults!(crate::FFT64AvxRayon, fft64);
+#[cfg(feature = "enable-rayon")]
+poulpy_cpu_ref::impl_cpu_core_defaults!(crate::FFT64CIAvxRayon, fft64);
+#[cfg(feature = "enable-rayon")]
+poulpy_cpu_ref::impl_cpu_core_defaults!(crate::NTT4x30AvxRayon, ntt4x30);
+#[cfg(feature = "enable-rayon")]
+poulpy_cpu_ref::impl_cpu_core_defaults!(crate::NTT4x30CIAvxRayon, ntt4x30);