+++ poulpy-cpu-avx512/src/ntt4x30_avx512/rayon.rs
@@ -774,6 +774,7 @@ unsafe impl<R: CpuRing> poulpy_core::oep::GGLWEProductDigitsStridedImpl for NTT4
     }
 }
 
+#[cfg(feature = "enable-ckks")]
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn vmp_apply_digits_strided_known_zero_prefix<R: CpuRing>(
     module: &Module<NTT4x30Avx512RayonBackend<R>>,